refactor: move parser-related serde code into parsers
